requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
django-cors-headers